+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-## [0.1] - 26th April 2019
+## [0.1.1] - 29th April 2019
+
+* setTextSize added
+* writeAll added
+* Breaking change : NetworkPOSPrinterUnderline.single_weight => NetworkPOSPrinterUnderline.single
+* Breaking change : NetworkPOSPrinterUnderline.double_weight => NetworkPOSPrinterUnderline.double
+
+## [0.1.0] - 26th April 2019
 
 Final version and example